+++ bedrock/extract/allocation/bea.py
@@ -11,7 +11,7 @@ from bedrock.transform.eeio.derived_2017 import (
     derive_2017_Y_personal_consumption_expenditure_usa,
 )
 from bedrock.utils.config.usa_config import get_usa_config
-from bedrock.utils.io.extract_input_local import local_extract_input_dir
+from bedrock.utils.io.local_extract_input_data import local_extract_input_dir
 from bedrock.utils.io.gcp import load_from_gcs
 from bedrock.utils.io.gcp_paths import gcs_extract_input_path
 from bedrock.utils.taxonomy.bea.ceda_v7 import CEDA_V7_SECTORS
