+++ bedrock/extract/allocation/epa.py
@@ -16,7 +16,7 @@ from bedrock.extract.allocation.epa_constants import (
 from bedrock.transform.allocation.utils import parse_index_with_aggregates
 from bedrock.utils.config.usa_config import get_usa_config
 from bedrock.utils.emissions.gwp import derive_ar5_to_ar6_multiplier
-from bedrock.utils.io.extract_input_local import local_dir_for_gcs_sub_bucket
+from bedrock.utils.io.local_extract_input_data import local_dir_for_gcs_sub_bucket
 from bedrock.utils.io.gcp import load_from_gcs
 from bedrock.utils.io.gcp_paths import gcs_extract_input_path
