+++ bedrock/extract/usda/USDA_CoA.py
@@ -19,7 +19,7 @@ from requests import Response
 from bedrock.transform.flowbyfunctions import assign_fips_location_system
 from bedrock.utils.config.common import WITHDRAWN_KEYWORD
 from bedrock.utils.config.schema import flow_by_activity_fields
-from bedrock.utils.io.extract_input_local import load_local_extract_input_dir
+from bedrock.utils.io.local_extract_input_data import load_local_extract_input_dir
 from bedrock.utils.io.gcp import load_from_gcs
 from bedrock.utils.io.gcp_paths import gcs_extract_input_sub_bucket_from_kwargs
 from bedrock.utils.mapping.location import US_FIPS, abbrev_us_state, to_ndigit_str
