+++ bedrock/utils/io/local_extract_input_data.py
